Домашна 3/on-chain/requirements.txt
python-dotenv
requests
pandas
sanpy
SQLAlchemy
psycopg2-binary
